refactor: extract subcontract filter helpers
