+++ app/subcontracts/page.tsx
@@ -1,9 +1,14 @@
 'use client';
 
-import { useState, useEffect, useMemo, useRef, Suspense, type CSSProperties, type ReactNode } from 'react';
-import { createPortal } from 'react-dom';
+import { useState, useEffect, useMemo, Suspense, type CSSProperties, type ReactNode } from 'react';
 import Link from 'next/link';
 import { useSearchParams, useRouter } from 'next/navigation';
+import { FilterRow } from '@/components/filters/FilterRow';
+import { FilterTextInput } from '@/components/filters/FilterTextInput';
+import { MinMaxInput } from '@/components/filters/MinMaxInput';
+import { MultiSelectDropdown } from '@/components/filters/MultiSelectDropdown';
+import { formatYen, parseAmountToYen } from '@/app/lib/format/yen';
+import { accountCategoryLabel, bureauLeaf } from '@/app/lib/subcontracts/labels';
 import type { SubcontractGraph } from '@/types/subcontract';
 
 type SortKey =
@@ -40,283 +45,6 @@ const STRING_SORT_KEYS: ReadonlySet<SortKey> = new Set<SortKey>([
 type SortDir = 'asc' | 'desc';
 
 const PAGE_SIZE = 50;
-
-function formatYen(v: number): string {
-  if (v >= 1e12) return `${(v / 1e12).toFixed(2)}兆円`;
-  if (v >= 1e10) return `${Math.round(v / 1e8).toLocaleString()}億円`;
-  if (v >= 1e8) return `${(v / 1e8).toFixed(2)}億円`;
-  if (v >= 1e4) return `${Math.round(v / 1e4).toLocaleString()}万円`;
-  return `${Math.round(v).toLocaleString()}円`;
-}
-
-/** sankey-svg と同じ会計区分ラベル表記 */
-function accountCategoryLabel(cat: string): string {
-  if (cat === '一般会計+特別会計') return '一般・特別';
-  if (cat === '一般会計') return '一般会計';
-  if (cat === '特別会計') return '特別会計';
-  return cat;
-}
-
-/** 担当組織の末端要素を取り出す */
-function bureauLeaf(bureau: string): string {
-  if (!bureau) return '';
-  const parts = bureau.split(' / ');
-  return parts[parts.length - 1] ?? '';
-}
-
-/** "1.26億", "100億", "1兆", "1兆2000億", "456789" などを1円単位の数値に変換。解析失敗時 null */
-function parseAmountToYen(input: string): number | null {
-  if (!input) return null;
-  const trimmed = input.trim().replace(/[,，\s]/g, '');
-  if (!trimmed) return null;
-  const combo = trimmed.match(/^([\d.]+)兆([\d.]+)億?$/);
-  if (combo) {
-    const cho = parseFloat(combo[1]);
-    const oku = parseFloat(combo[2]);
-    if (!isNaN(cho) && !isNaN(oku)) return (cho * 10000 + oku) * 1e8;
-  }
-  const m = trimmed.match(/^([\d.]+)\s*(兆|億|万|千)?円?$/);
-  if (!m) return null;
-  const v = parseFloat(m[1]);
-  if (isNaN(v)) return null;
-  switch (m[2]) {
-    case '兆': return v * 1e12;
-    case '億': return v * 1e8;
-    case '万': return v * 1e4;
-    case '千': return v * 1e3;
-    default: return v;
-  }
-}
-
-/** sankey-svg 風の複数選択ドロップダウン */
-interface MultiSelectDropdownProps {
-  options: string[];
-  selected: string[];
-  onChange: (next: string[]) => void;
-  allLabel: string;
-  placeholder?: string;
-  minWidth?: number;
-}
-
-/** sankey-svg 風の label + control + ✕ 行 */
-function FilterRow({ label, children }: { label: string; children: ReactNode }) {
-  return (
-    <div style={{ display: 'flex', alignItems: 'center', gap: 6 }}>
-      <span style={{ fontSize: 11, color: '#555', width: 40, flexShrink: 0, fontWeight: 600 }}>{label}</span>
-      <div style={{ flex: 1, minWidth: 0, display: 'flex', alignItems: 'center', gap: 4 }}>
-        {children}
-      </div>
-    </div>
-  );
-}
-
-/** 部分一致テキスト入力（クリアボタン付き） */
-function FilterTextInput({ value, onChange, placeholder }: { value: string; onChange: (v: string) => void; placeholder?: string }) {
-  return (
-    <div style={{ position: 'relative', flex: 1, minWidth: 0 }}>
-      <input
-        type="text"
-        value={value}
-        onChange={(e) => onChange(e.target.value)}
-        placeholder={placeholder}
-        style={{
-          width: '100%',
-          boxSizing: 'border-box',
-          fontSize: 12,
-          border: '1px solid #ddd',
-          borderRadius: 4,
-          padding: '3px 22px 3px 6px',
-          background: '#fafafa',
-          color: '#333',
-          outline: 'none',
-        }}
-      />
-      {value && (
-        <button
-          type="button"
-          onClick={() => onChange('')}
-          aria-label="クリア"
-          style={{ position: 'absolute', right: 4, top: '50%', transform: 'translateY(-50%)', background: 'none', border: 'none', color: '#bbb', cursor: 'pointer', padding: 2, fontSize: 11 }}
-        >
-          ✕
-        </button>
-      )}
-    </div>
-  );
-}
-
-/** Min-Max 金額入力ペア */
-function MinMaxInput({
-  minVal, maxVal, onMinChange, onMaxChange,
-}: {
-  minVal: string; maxVal: string;
-  onMinChange: (v: string) => void; onMaxChange: (v: string) => void;
-}) {
-  const minOk = !minVal || parseAmountToYen(minVal) !== null;
-  const maxOk = !maxVal || parseAmountToYen(maxVal) !== null;
-  const inputStyle = (ok: boolean): CSSProperties => ({
-    flex: 1,
-    minWidth: 0,
-    fontSize: 12,
-    border: `1px solid ${ok ? '#ddd' : '#e53935'}`,
-    borderRadius: 4,
-    padding: '3px 6px',
-    background: '#fafafa',
-    color: '#333',
-    outline: 'none',
-  });
-  return (
-    <>
-      <input
-        type="text"
-        value={minVal}
-        onChange={(e) => onMinChange(e.target.value)}
-        placeholder="下限"
-        title="下限 (例: 100億, 1兆)"
-        style={inputStyle(minOk)}
-      />
-      <span style={{ color: '#aaa', fontSize: 11 }}>〜</span>
-      <input
-        type="text"
-        value={maxVal}
-        onChange={(e) => onMaxChange(e.target.value)}
-        placeholder="上限"
-        title="上限 (例: 1兆, 5000億)"
-        style={inputStyle(maxOk)}
-      />
-      {(minVal || maxVal) && (
-        <button
-          type="button"
-          onClick={() => { onMinChange(''); onMaxChange(''); }}
-          aria-label="クリア"
-          style={{ background: 'none', border: 'none', color: '#bbb', cursor: 'pointer', padding: 2, fontSize: 11, flexShrink: 0 }}
-        >
-          ✕
-        </button>
-      )}
-    </>
-  );
-}
-
-function MultiSelectDropdown({ options, selected, onChange, allLabel, minWidth = 160 }: MultiSelectDropdownProps) {
-  const [open, setOpen] = useState(false);
-  const [rect, setRect] = useState<{ top: number; left: number; width: number; maxHeight: number } | null>(null);
-  const buttonRef = useRef<HTMLButtonElement>(null);
-  const dropdownRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!open) return;
-    function onPointerDown(e: MouseEvent) {
-      if (dropdownRef.current?.contains(e.target as Node)) return;
-      if (buttonRef.current?.contains(e.target as Node)) return;
-      setOpen(false);
-    }
-    document.addEventListener('mousedown', onPointerDown);
-    return () => document.removeEventListener('mousedown', onPointerDown);
-  }, [open]);
-
-  const allSelected = selected.length === 0;
-  const label = allSelected
-    ? allLabel
-    : selected.length === 1
-      ? selected[0]
-      : `選択中 (${selected.length}/${options.length})`;
-
-  return (
-    <div style={{ position: 'relative', minWidth, flex: 1 }}>
-      <button
-        type="button"
-        ref={buttonRef}
-        onClick={() => {
-          if (buttonRef.current) {
-            const r = buttonRef.current.getBoundingClientRect();
-            setRect({
-              top: r.bottom + 2,
-              left: r.left,
-              width: Math.max(r.width, 200),
-              maxHeight: Math.max(160, window.innerHeight - r.bottom - 24),
-            });
-          }
-          setOpen((v) => !v);
-        }}
-        style={{
-          width: '100%',
-          fontSize: 12,
-          border: '1px solid #ddd',
-          borderRadius: 4,
-          padding: '3px 22px 3px 6px',
-          background: '#fafafa',
-          color: allSelected ? '#aaa' : '#333',
-          cursor: 'pointer',
-          textAlign: 'left',
-          overflow: 'hidden',
-          textOverflow: 'ellipsis',
-          whiteSpace: 'nowrap',
-          outline: 'none',
-        }}
-      >
-        {label}
-      </button>
-      <svg xmlns="http://www.w3.org/2000/svg" height="12" width="12" viewBox="0 0 24 24" fill="#aaa"
-        style={{ position: 'absolute', right: 6, top: '50%', transform: open ? 'translateY(-50%) rotate(180deg)' : 'translateY(-50%)', transition: 'transform 0.15s', pointerEvents: 'none' }}>
-        <path d="M7 10l5 5 5-5z"/>
-      </svg>
-      {!allSelected && (
-        <button
-          type="button"
-          onClick={() => onChange([])}
-          aria-label="クリア"
-          style={{ position: 'absolute', right: 22, top: '50%', transform: 'translateY(-50%)', background: 'none', border: 'none', color: '#bbb', cursor: 'pointer', padding: 2, fontSize: 11 }}
-        >
-          ✕
-        </button>
-      )}
-      {open && rect && typeof document !== 'undefined' && createPortal(
-        <div
-          ref={dropdownRef}
-          style={{
-            position: 'fixed',
-            top: rect.top,
-            left: rect.left,
-            width: rect.width,
-            maxHeight: rect.maxHeight,
-            overflowY: 'auto',
-            zIndex: 9999,
-            background: '#fff',
-            border: '1px solid #ddd',
-            borderRadius: 4,
-            boxShadow: '0 4px 12px rgba(0,0,0,0.12)',
-          }}
-          onMouseDown={(e) => e.stopPropagation()}
-        >
-          <label style={{ display: 'flex', alignItems: 'center', gap: 6, padding: '5px 8px', cursor: 'pointer', borderBottom: '1px solid #f0f0f0', fontWeight: 600 }}>
-            <input
-              type="checkbox"
-              checked={allSelected}
-              onChange={() => onChange([])}
-              style={{ width: 12, height: 12 }}
-            />
-            <span style={{ fontSize: 12, color: '#333' }}>すべて解除</span>
-          </label>
-          {options.map((opt) => (
-            <label key={opt} style={{ display: 'flex', alignItems: 'center', gap: 6, padding: '4px 8px', cursor: 'pointer' }}>
-              <input
-                type="checkbox"
-                checked={selected.includes(opt)}
-                onChange={() =>
-                  onChange(selected.includes(opt) ? selected.filter((x) => x !== opt) : [...selected, opt])
-                }
-                style={{ width: 12, height: 12 }}
-              />
-              <span style={{ fontSize: 12, color: '#333' }}>{opt}</span>
-            </label>
-          ))}
-        </div>,
-        document.body
-      )}
-    </div>
-  );
-}
 
 function SubcontractsPageInner() {
   const searchParams = useSearchParams();
